refactor(General): :art: Implement camelCase and diffferent subclasses for handeling vm, backups and monitoring
THis commit introduces the use of camelCase instead of the previous used snake_case for better readability. I also changed some function names, so instead of backupDelete -> deleteBackup. This also implements different subclassess in the base.py for VM, Backup and Monitoring stuff, this should help the readability of the code and its usage when using and ide ig.

Implementation of camelCase instead of snake_case  and subclasses might lead to uncompatible code of old codebases